added make targets for building primme as a shared library
The primme build system currently does not support building primme as a
shared library. This is insofar unfortunate as the LGPL license does not
permit static linking into proprietary software. I have added three
build targets (solib, solibd, solibz) that build shared versions of the
primme library to the makefile. This works by taking the static
library and dumping its contents into a shared object, so the changes
are quite minimal. Note that -fPIC has to be enabled in the Make_flags
file if a shared library is to be built.